+++ lib/aua-mite/agents/mighty_mite.rb
@@ -1,7 +1,7 @@
 module Aua::Agents::MightyMite
 
   def self.extend?(agent)
-    agent.app == "MightyMite"
+    agent.app == "MightyMite" || agent.app == "mighty-mite"
   end
 
   def type
@@ -13,6 +13,6 @@ module Aua::Agents::MightyMite
   end
 
   def version
-    @version ||= parts[1] && parts[1][0]
+    @version ||= app == "MightyMite" ? parts[1] && parts[1][0] : version_of("mighty-mite")
   end
 end